refactor: style 전체 분리
