+++ src/components/homePage/SummaryHome.tsx
@@ -1,11 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
-import HomeSummaryBg from '../images/HomeSummaryBg.png';
 import axios from 'axios';
 import NotSummaryTeample from 'components/teampleHomePage/nothing/NotSummaryTeample';
 import { useRecoilState } from 'recoil';
 import { profileImgState, isLoading4State } from 'state';
 import { baseURL } from 'api/client';
+import * as Style from '../../css/HomePage/SummaryHomeStyle';
 
 const SummaryHome = () => {
   const now = new Date();
@@ -62,13 +61,13 @@ const SummaryHome = () => {
       {allNum === 0 ? (
         <NotSummaryTeample />
       ) : (
-        <SummaryContainer>
-          <DateContainer>
+        <Style.SummaryContainer>
+          <Style.DateContainer>
             {year}년 {month}월 {date}일 ({week})
-          </DateContainer>
-          <RemainContainer>
-            <RemainBox>
-              <Big>
+          </Style.DateContainer>
+          <Style.RemainContainer>
+            <Style.RemainBox>
+              <Style.Big>
                 <span>
                   {username}님의
                   <br />
@@ -83,134 +82,28 @@ const SummaryHome = () => {
                     <br />
                   </span>
                 </div>
-              </Big>
-              <Small>
-                <Text>남은 시간 힘내서 달려보아요</Text>
-                <Percent>{remainPercent}%</Percent>
-              </Small>
-            </RemainBox>
-          </RemainContainer>
-          <BarContainer>
+              </Style.Big>
+              <Style.Small>
+                <Style.Text>남은 시간 힘내서 달려보아요</Style.Text>
+                <Style.Percent>{remainPercent}%</Style.Percent>
+              </Style.Small>
+            </Style.RemainBox>
+          </Style.RemainContainer>
+          <Style.BarContainer>
             <ul>
               <li>
-                <Bar
+                <Style.Bar
                   className="css-progressbar"
                   remainPercent={remainPercent}
                   profileImage={profileImg}
                 />
               </li>
             </ul>
-          </BarContainer>
-        </SummaryContainer>
+          </Style.BarContainer>
+        </Style.SummaryContainer>
       )}
     </>
   );
 };
-
-const SummaryContainer = styled.div`
-  width: 87.5vw;
-  height: 30vh;
-  position: relative;
-  background-color: #f9fafd;
-  background-image: url(${HomeSummaryBg});
-  background-size: contain;
-  background-repeat: no-repeat;
-`;
-
-const DateContainer = styled.div`
-  position: absolute;
-  left: 2.81vw;
-  top: 3.33vh;
-  font-size: 0.83vw;
-  line-height: 100%;
-`;
-
-const RemainContainer = styled.div``;
-
-const RemainBox = styled.div`
-  position: absolute;
-  top: 6.85185vh;
-  left: 2.81vw;
-`;
-
-const Big = styled.div`
-  font-size: 1.67vw;
-  line-height: 100%;
-  font-weight: 500;
-`;
-
-const Small = styled.div`
-  margin-top: 1.851852vh;
-`;
-
-const Text = styled.span`
-  font-size: 0.9375vw;
-  font-weight: 400;
-`;
-
-const Percent = styled.span`
-  position: absolute;
-  left: 43.75vw;
-  font-size: 1.25vw;
-`;
-
-const BarContainer = styled.div`
-  position: absolute;
-  top: 21.2963vh;
-  left: 4vw;
-  ul {
-    position: relative;
-    padding: 0;
-    list-style: none;
-    width: 46.25vw;
-  }
-
-  li {
-    background-color: #ececec;
-    height: 16px;
-    border-radius: 46px;
-  }
-`;
-
-interface IBar {
-  remainPercent: number;
-  profileImage: string;
-}
-
-const Bar = styled.span<IBar>`
-  position: absolute;
-  border-radius: 46px;
-  background-color: #487aff;
-  height: 16px;
-  width: ${(props) => props.remainPercent}%;
-  -webkit-animation: css-progressbar 2s ease-out;
-  animation: css-progressbar 2s ease-out;
-
-  &::after {
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-    content: '';
-    width: 40px;
-    height: 40px;
-    position: absolute;
-    top: -10px;
-    right: -1.5vw;
-    border: 1.5px solid #487aff;
-    border-radius: 54px;
-    background-image: url(${(props) => require(`../images/profile/proImageU` + props.profileImage+ `.png`)});
-    background-size: contain;
-    background-repeat: no-repeat;
-  }
-
-  @keyframes css-progressbar {
-    0% {
-      width: 0px;
-    }
-
-    100% {
-      width: ${(props) => props.remainPercent}%;
-    }
-  }
-`;
 
 export default SummaryHome;
