+++ src/components/moreInfoPage/InfoInput.tsx
@@ -1,6 +1,4 @@
-import React, { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import axios from 'axios';
+import React, { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { useNavigate } from 'react-router';
 import {
@@ -10,8 +8,8 @@ import {
   jwtAccessTokenState,
   jwtRefreshTokenState,
 } from 'state';
-import { baseURL } from 'api/client';
 import authAPI from 'api/authAPI';
+import * as Style from '../../css/MoreInfoPage/InfoInputStyle';
 
 const InfoInput = () => {
   const [name, setName] = useState('');
@@ -50,7 +48,16 @@ const InfoInput = () => {
     //     profileImage: `proImageU${randomNum}`,
     //   },
     // })
-    authAPI.postCreate(idToken,kakaoAccessToken,kakaoRefreshToken,name, school, major, randomNum)
+    authAPI
+      .postCreate(
+        idToken,
+        kakaoAccessToken,
+        kakaoRefreshToken,
+        name,
+        school,
+        major,
+        randomNum,
+      )
       .then((response) => {
         setjwtAccessToken(response.data.data.jwtAccessToken);
         setjwtRefreshToken(response.data.data.jwtRefreshToken);
@@ -72,21 +79,21 @@ const InfoInput = () => {
 
   return (
     <div>
-      <InputContainer>
-        <InputBox>
-          <PlaceHolder>이름</PlaceHolder>
-          <Input onChange={onChangeName} />
-        </InputBox>
-        <InputBox>
-          <PlaceHolder>학교</PlaceHolder>
-          <Input onChange={onChangeSchool} />
-        </InputBox>
-        <InputBox>
-          <PlaceHolder>학과</PlaceHolder>
-          <Input onChange={onChangeMajor} />
-        </InputBox>
-      </InputContainer>
-      <Btn
+      <Style.InputContainer>
+        <Style.InputBox>
+          <Style.PlaceHolder>이름</Style.PlaceHolder>
+          <Style.Input onChange={onChangeName} />
+        </Style.InputBox>
+        <Style.InputBox>
+          <Style.PlaceHolder>학교</Style.PlaceHolder>
+          <Style.Input onChange={onChangeSchool} />
+        </Style.InputBox>
+        <Style.InputBox>
+          <Style.PlaceHolder>학과</Style.PlaceHolder>
+          <Style.Input onChange={onChangeMajor} />
+        </Style.InputBox>
+      </Style.InputContainer>
+      <Style.Btn
         onClick={postAuthInfoAPI}
         disabled={
           name.length !== 0 && school.length !== 0 && major.length !== 0
@@ -94,49 +101,10 @@ const InfoInput = () => {
             : true
         }
       >
-        <Text>팀쁠 시작하기</Text>
-      </Btn>
+        <Style.Text>팀쁠 시작하기</Style.Text>
+      </Style.Btn>
     </div>
   );
 };
-const InputBox = styled.div`
-  width: 372px;
-  height: 54px;
-  background-color: #ffffff;
-  border-radius: 8px;
-  font-size: 20px;
-  line-height: 100%;
-  color: #383838;
-  padding: 0px 20px;
-  display: flex;
-  align-items: center;
-  margin-bottom: 16px;
-`;
-const InputContainer = styled.div`
-  margin-bottom: 40px;
-`;
-const PlaceHolder = styled.span`
-  margin-right: 24px;
-  color: #707070;
-  font-size: 20px;
-`;
-const Input = styled.input`
-  border: none;
-  font-size: 20px;
-  width: 270px;
-`;
-const Btn = styled.button`
-  width: 372px;
-  height: 54px;
-  background-color: #487aff;
-  border-radius: 8px;
-  &:disabled {
-    background-color: #d4e4ff;
-  }
-`;
-const Text = styled.span`
-  font-size: 20px;
-  line-height: 100%;
-  color: #ffffff; ;
-`;
+
 export default InfoInput;
