+++ src/components/teampleDetailPage/detailBox.tsx
@@ -1,4 +1,3 @@
-import styled from 'styled-components';
 import React, { useEffect, useState, useRef } from 'react';
 import vector from '../images/Vector.png';
 import more from '../images/Group 419.png';
@@ -7,10 +6,8 @@ import finBtn from '../images/Group 879.png';
 import download from '../images/DownloadSimple.png';
 import trash from '../images/Trash.png';
 import deleteBtn from '../images/delete.png';
-import ellipse from '../images/Ellipse 1.png';
 import Send from '../images/send.png';
 import axios from 'axios';
-import { detailInfo } from 'interfaces/taskType';
 import { userInfo } from 'interfaces/userType';
 import S3 from 'react-aws-s3-typescript';
 import { config } from 'config';
@@ -32,399 +29,7 @@ import filesApi from 'api/fileAPI';
 import feedbackAPI from 'api/feedbackAPI';
 import taskAPI from 'api/taskAPI';
 import userAPI from 'api/userAPI';
-
-const DetailContainer = styled.div`
-  width: 52.0833vw;
-  height: 93.333vh;
-  display: flex;
-  flex-direction: column;
-  z-index: 1;
-  margin-left: auto;
-  margin-right: auto;
-
-  .headerBtns {
-    display: flex;
-    justify-content: space-between;
-    z-index: 1000;
-  }
-
-  .back {
-    width: 1.667vw;
-    height: 2.963vh;
-    margin-top: 1.7vh;
-  }
-
-  .more {
-    margin-top: 3.0556vh;
-    width: 0.26042vw;
-    height: 1.9444vh;
-    margin-right: 0.9896vw;
-  }
-
-  .more:hover,
-  .back:hover {
-    cursor: grab;
-  }
-
-  img {
-    max-width: 100%;
-    max-height: 100%;
-  }
-
-  .toDoInfoBox {
-    width: 100%;
-    height: 20.1852vh;
-    margin-top: 2.963vh;
-    margin-bottom: 7.037vh;
-  }
-
-  .step {
-    display: flex;
-  }
-
-  .stepRound {
-    width: 2.65625vw;
-    height: 2.222vh;
-    border: 1px solid #487aff;
-    border-radius: 20px;
-    display: flex;
-  }
-
-  .stepText {
-    font-weight: 500;
-    font-size: 0.625vw;
-    line-height: 100%;
-    color: #487aff;
-    margin: auto;
-  }
-
-  .stepName {
-    font-weight: 500;
-    font-size: 0.9375vw;
-    line-height: 100%;
-    color: #487aff;
-    margin-left: 0.41667vw;
-    margin-top: auto;
-    margin-bottom: auto;
-  }
-
-  .subName {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 2.59259vh;
-  }
-
-  .taskName {
-    font-weight: 600;
-    font-size: 1.45833vw;
-    line-height: 100%;
-    color: #383838;
-  }
-
-  .finBtn {
-    width: 7.08333vw;
-    height: 3.333vh;
-    margin-right: 1.04167vw;
-  }
-
-  .finBtn:hover {
-    cursor: grab;
-  }
-
-  .subInfo {
-    display: flex;
-    flex-direction: column;
-    margin-top: 3.425925vh;
-  }
-
-  .manager,
-  .date,
-  .state {
-    font-weight: 500;
-    font-size: 0.9375vw;
-    line-height: 100%;
-    color: #707070;
-    margin-bottom: 2.31481vh;
-  }
-  .managerInput {
-    margin-left: 2.70833vw;
-    font-weight: 400;
-    font-size: 1.041667vw;
-  }
-
-  .dateInput {
-    margin-left: 3.489583vw;
-    font-weight: 400;
-    font-size: 1.04167vw;
-  }
-
-  .stateInput {
-    margin-left: 1.6667vw;
-    font-weight: 400;
-    font-size: 1.04167vw;
-  }
-
-  .top {
-    margin-bottom: 6.574074vh;
-  }
-
-  .file {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .fileText {
-    font-weight: 500;
-    font-size: 1.04167vw;
-    line-height: 100%;
-    color: #383838;
-    margin-left: 0.3125vw;
-    margin-top: auto;
-    margin-bottom: auto;
-  }
-
-  .addFile {
-    width: 6.3vw;
-    height: 3.5vh;
-    margin-left: 24px;
-    border: 1px solid #d5dbee;
-    border-radius: 8px;
-    color: #707070;
-    font-size: 0.625vw;
-  }
-
-  .addFile:hover {
-    cursor: grab;
-  }
-
-  .files {
-    margin-top: 1.85185vh;
-    flex-wrap: nowrap;
-    overflow: auto;
-    display: flex;
-  }
-  .uploadDate {
-    width: 91px;
-    height: 12px;
-    left: 614px;
-    top: 565px;
-  }
-
-  .fileCard {
-    width: 14.1667vw;
-    height: 8.5185vh;
-    background: #ffffff;
-    border-radius: 8px;
-    border: 1px solid #d5dbee;
-    margin-right: 1.45833vw;
-    min-width: 14.1667vw;
-  }
-
-  .fileName {
-    display: flex;
-    margin-top: 1.85195vh;
-    margin-left: 1.041667vw;
-  }
-
-  .nameText {
-    font-weight: 500;
-    font-size: 0.8333vw;
-    line-height: 100%;
-    text-align: center;
-    color: #707070;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-  }
-
-  .icons {
-    margin-left: auto;
-    margin-right: 2.3vw;
-    width: 1.3vw;
-    display: flex;
-  }
-
-  .download {
-    margin-right: 0.41667vw;
-    width: 3vw;
-    object-fit: cover;
-  }
-
-  .trash {
-    width: 3vw;
-    object-fit: cover;
-  }
-
-  .download:hover,
-  .trash:hover {
-    cursor: grab;
-  }
-
-  .fileInfo {
-    font-weight: 400;
-    font-size: 0.625vw;
-    line-height: 100%;
-    color: #a7a7a7;
-    margin-left: 1.041667vw;
-    display: flex;
-    margin-top: 1.85185vh;
-  }
-
-  .fileSize {
-    margin-left: auto;
-    margin-right: 1.041667vw;
-  }
-
-  .mid {
-    margin-bottom: 6.6667vh;
-  }
-
-  .feedText {
-    font-weight: 500;
-    font-size: 1.041667vw;
-    line-height: 100%;
-    color: #383838;
-    margin-left: 0.3125vw;
-  }
-
-  .addFeed {
-    display: flex;
-    margin-top: 2.222vh;
-  }
-
-  .profileImg {
-    width: 2.5vw;
-    height: 4.444vh;
-    margin-left: 0.625vw;
-  }
-
-  .feedInput {
-    width: 42.916667vw;
-    height: 4.444vh;
-    margin-top: auto;
-    margin-bottom: auto;
-    margin-left: 0.8333vw;
-    border: 0 solid black;
-    padding: 0;
-    background: rgba(237, 239, 246, 0.5);
-    border-radius: 8px;
-    padding-left: 0.83333vw;
-
-    ::-webkit-input-placeholder {
-      margin-left: 16px;
-      color: '#CCCCCC';
-      font-size: 0.83333vw;
-      margin-left: 16px;
-    }
-  }
-
-  .feedInput:focus {
-    outline: none;
-    border: solid 1px #487aff;
-  }
-
-  .inputBox {
-    position: relative;
-    ::placeholder {
-      color: #cccccc;
-    }
-  }
-
-  .send {
-    position: absolute;
-    top: 1.2vh;
-    left: 44.7vw;
-    border: none;
-    background-color: transparent;
-    color: #a7a7a7;
-    width: 17px;
-    height: 17px;
-  }
-
-  .send:hover {
-    cursor: grab;
-  }
-
-  .feedbacks {
-    display: flex;
-    flex-direction: column;
-    margin-top: 3.333vh;
-  }
-
-  .feedBox {
-    width: 50.625vw;
-    height: 7.314814vh;
-    display: flex;
-    margin-bottom: 1.48148vh;
-    overflow-y: auto;
-    overflow-x: hidden;
-  }
-
-  .fullFeed {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .feedInfo {
-    display: flex;
-    margin-left: 0.83333vw;
-    width: 100%;
-  }
-
-  .feedName {
-    font-weight: 500;
-    font-size: 0.9375vw;
-    line-height: 100%;
-    color: #383838;
-  }
-
-  .feedDate {
-    font-weight: 500;
-    font-size: 0.72917vw;
-    line-height: 100%;
-    color: #a7a7a7;
-    margin-left: 0.625vw;
-    margin-top: auto;
-    margin-bottom: auto;
-  }
-
-  .plusBtn {
-    height: 0.7604vw;
-    width: 2.944vh;
-    margin-left: auto;
-    margin-right: 1.041667vw;
-    margin-top: auto;
-  }
-
-  .plusBtn:hover {
-    cursor: grab;
-  }
-
-  .feedContent {
-    font-weight: 400;
-    font-size: 0.833vw;
-    line-height: 100%;
-    color: #383838;
-    margin-left: 0.8333vw;
-    margin-top: 1.111vh;
-    max-width: 70%;
-  }
-`;
-
-const Container = styled.div`
-  width: 87.5vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-`;
-
-const ModalContainer = styled.div`
-  position: absolute;
-  top: 40px;
-  left: -140px;
-  z-index: 1000;
-`;
+import * as Style from '../../css/TeampleDetailPage/DetailBoxStyle';
 
 const DetailBox = () => {
   const token = localStorage.getItem('jwt_accessToken');
@@ -495,7 +100,8 @@ const DetailBox = () => {
   };
 
   const getUser = async () => {
-    userAPI.getUserProfile()
+    userAPI
+      .getUserProfile()
       .then((res) => {
         setUser(res.data.data);
       })
@@ -506,14 +112,15 @@ const DetailBox = () => {
   //새로고침 시에도 taskId에 맞는 디테일 정보를 가져와야 해서 양쪽에서 모두 get함수를..;;
 
   const postFeedback = async () => {
-    feedbackAPI.post(taskId,addFeed)
-        .then(() => {
-          setAddFeed('');
-          location.reload();
-        })
-        .catch((e) => {
-          console.log(e);
-        });
+    feedbackAPI
+      .post(taskId, addFeed)
+      .then(() => {
+        setAddFeed('');
+        location.reload();
+      })
+      .catch((e) => {
+        console.log(e);
+      });
   };
   useEffect(() => {
     getUser();
@@ -524,7 +131,8 @@ const DetailBox = () => {
   }, [taskId]);
 
   useDidMountEffect(() => {
-      filesApi.postFile(file?.name,file?.size,fileLoc,taskId,teamid)
+    filesApi
+      .postFile(file?.name, file?.size, fileLoc, taskId, teamid)
       .then(() => {
         alert('파일 등록이 완료되었습니다.');
         location.reload();
@@ -535,17 +143,19 @@ const DetailBox = () => {
   }, [fileLoc]);
 
   const onChangeStatus = async () => {
-    taskAPI.toggle(taskId)
-    .then(() => {
-      location.reload();
-    })
-    .catch((e) => {
-      console.log(e);
-    });
+    taskAPI
+      .toggle(taskId)
+      .then(() => {
+        location.reload();
+      })
+      .catch((e) => {
+        console.log(e);
+      });
   };
 
   const onDeleteFeed = async (feedId: number) => {
-    feedbackAPI.delete(feedId)
+    feedbackAPI
+      .delete(feedId)
       .then(() => {
         location.reload();
       })
@@ -581,20 +191,21 @@ const DetailBox = () => {
   };
 
   const delTaskAPI = (fileId: number) => {
-    filesApi.delFileAPI(fileId)
+    filesApi
+      .delFileAPI(fileId)
       .then(() => {
-          location.reload();
-        })
+        location.reload();
+      })
       .catch((e) => {
-          console.log(e);
-        });
+        console.log(e);
+      });
   };
 
   const alertDelFile = (e: React.MouseEvent<HTMLElement>) => {
     confirmAlert({
       customUI: ({ onClose }) => {
         return (
-          <AlertFile>
+          <Style.AlertFile>
             <div className="alertBody">삭제하시면 복구할 수 없어요.</div>
             <div className="alertTitle">정말 파일을 삭제하시겠어요?</div>
             <div className="alertButtons">
@@ -610,7 +221,7 @@ const DetailBox = () => {
                 네,삭제할래요.
               </button>
             </div>
-          </AlertFile>
+          </Style.AlertFile>
         );
       },
     });
@@ -620,7 +231,7 @@ const DetailBox = () => {
     confirmAlert({
       customUI: ({ onClose }) => {
         return (
-          <AlertFeed>
+          <Style.AlertFeed>
             <div className="alertTitle">피드백을 삭제하시겠어요?</div>
             <div className="alertButtons">
               <button onClick={onClose} className="alertNo">
@@ -635,19 +246,19 @@ const DetailBox = () => {
                 네,삭제할래요.
               </button>
             </div>
-          </AlertFeed>
+          </Style.AlertFeed>
         );
       },
     });
   };
 
   return (
-    <Container>
-      <ModalContainer>
+    <Style.Container>
+      <Style.ModalContainer>
         {bigModal && <ModifyTask setBigModal={setBigModal} />}
-      </ModalContainer>
+      </Style.ModalContainer>
       {detail && (
-        <DetailContainer>
+        <Style.DetailContainer>
           <div className="headerBtns">
             <div className="back" onClick={() => navigate(-1)}>
               <img src={vector} />
@@ -658,9 +269,9 @@ const DetailBox = () => {
               style={{ position: 'relative' }}
             >
               <img src={more} />
-              <ModalContainer>
+              <Style.ModalContainer>
                 {smallModal && <MoreTeampleDetail />}
-              </ModalContainer>
+              </Style.ModalContainer>
             </div>
           </div>
           <div className="toDoInfoBox">
@@ -828,101 +439,11 @@ const DetailBox = () => {
               </div>
             )}
           </div>
-        </DetailContainer>
+        </Style.DetailContainer>
       )}
-    </Container>
+    </Style.Container>
   );
 };
-
-const AlertFile = styled.div`
-  width: 440px;
-  height: 168px;
-  border-radius: 12px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: relative;
-  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.1);
-  background-color: white;
-
-  .alertButtons {
-    position: absolute;
-    bottom: 34px;
-    right: 34px;
-    display: flex;
-    justify-content: right;
-  }
-
-  .alertBody {
-    position: absolute;
-    top: 40px;
-    font-size: 14px;
-    color: #707070;
-    font-weight: 400;
-  }
-
-  .alertTitle {
-    position: absolute;
-    font-size: 18px;
-    color: #383838;
-    font-weight: 600;
-    top: 66px;
-  }
-
-  .alertYes {
-    color: #487aff;
-    font-weight: 600;
-    font-size: 16px;
-  }
-
-  .alertNo {
-    color: #a7a7a7;
-    font-weight: 600;
-    font-size: 16px;
-    margin-right: 20px;
-  }
-`;
-
-const AlertFeed = styled.div`
-  width: 440px;
-  height: 144px;
-  border-radius: 12px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: relative;
-  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.15);
-  background-color: white;
-
-  .alertButtons {
-    position: absolute;
-    bottom: 34px;
-    right: 34px;
-    display: flex;
-    justify-content: right;
-  }
-
-  .alertTitle {
-    position: absolute;
-    font-size: 18px;
-    color: #383838;
-    font-weight: 600;
-    top: 40px;
-  }
-
-  .alertYes {
-    color: #487aff;
-    font-weight: 600;
-    font-size: 16px;
-  }
-
-  .alertNo {
-    color: #a7a7a7;
-    font-weight: 600;
-    font-size: 16px;
-    margin-right: 20px;
-  }
-`;
 
 export default DetailBox;
