+++ src/components/teampleHomePage/FileInfo.tsx
@@ -1,11 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import axios from 'axios';
 import { useRecoilState } from 'recoil';
 import { teamidState } from 'state';
-import { baseURL } from 'api/client';
 import fileAPI from 'api/fileAPI';
+import * as Style from '../../css/TeampleHomePage/FileInfoStyle';
 
 interface IFileMap {
   size: number;
@@ -15,124 +13,52 @@ const FileInfo = () => {
   const [fileNum, setFileNum] = useState(0);
   const [nowFileSize, setNowFileSize] = useState(0);
   const [remainPercent, setRemainPercent] = useState((nowFileSize / 5) * 100);
-  const [count, setCount] = useState(0);
   let fsize = 0;
   const [teamid] = useRecoilState(teamidState);
-  const token = localStorage.getItem('jwt_accessToken');
 
   useEffect(() => {
-    fileAPI.getFiles(teamid)
-        .then((res) => {
-          setFileNum(res.data.data.length);
-          {
-            res.data.data.map((file: IFileMap) => {
-              fsize += file.size;
-              setNowFileSize(fsize);
-            });
-          }
-        })
-        .catch((e) => {
-          console.log(e);
-        });
+    fileAPI
+      .getFiles(teamid)
+      .then((res) => {
+        setFileNum(res.data.data.length);
+        {
+          res.data.data.map((file: IFileMap) => {
+            fsize += file.size;
+            setNowFileSize(fsize);
+          });
+        }
+      })
+      .catch((e) => {
+        console.log(e);
+      });
   }, [teamid]);
 
   return (
-    <FileInfoContainer>
+    <Style.FileInfoContainer>
       <Link style={{ textDecoration: 'none', color: 'black' }} to="/file">
-        <FileInfoBox>
-          <Title>공유 파일함</Title>
-          <FileNum>
+        <Style.FileInfoBox>
+          <Style.Title>공유 파일함</Style.Title>
+          <Style.FileNum>
             <span style={{ fontSize: '1.25vw' }}>{fileNum}</span>
             <span style={{ color: '#707070' }}>&nbsp;파일</span>
-          </FileNum>
-          <FileSize>{(nowFileSize / (1024 * 1024)).toFixed(1)}MB/5GB</FileSize>
-          <BarContainer>
+          </Style.FileNum>
+          <Style.FileSize>
+            {(nowFileSize / (1024 * 1024)).toFixed(1)}MB/5GB
+          </Style.FileSize>
+          <Style.BarContainer>
             <ul>
               <li>
-                <Bar
+                <Style.Bar
                   className="file-progressbar"
                   remainPercent={remainPercent}
                 />
               </li>
             </ul>
-          </BarContainer>
-        </FileInfoBox>
+          </Style.BarContainer>
+        </Style.FileInfoBox>
       </Link>
-    </FileInfoContainer>
+    </Style.FileInfoContainer>
   );
 };
-
-const FileInfoContainer = styled.div`
-  margin-top: 7.592vh;
-`;
-
-const FileInfoBox = styled.div`
-  width: 19.375vw;
-  height: 13.5vh;
-  background: #fffce3;
-  border-radius: 12px;
-  position: relative;
-`;
-
-const Title = styled.div`
-  font-size: 1.2vw;
-  position: absolute;
-  top: 2.2222vh;
-  left: 1.25vw;
-`;
-
-const FileNum = styled.div`
-  position: absolute;
-  top: 2.2222vh;
-  left: 15.1vw;
-  font-size: 1.2vw;
-`;
-
-const FileSize = styled.div`
-  color: #707070;
-  position: absolute;
-  top: 8.1041vh;
-  right: 1.25vw;
-  height: 1.481481vh;
-  font-size: 1vw;
-`;
-
-const BarContainer = styled.div`
-  position: absolute;
-  bottom: 22px;
-  height: 0.8vh;
-  ul {
-    position: relative;
-    padding: 0;
-    list-style: none;
-    width: 19.375vw;
-  }
-
-  li {
-    background-color: #fff8b7;
-    height: 1.481481vh;
-    border-radius: 46px;
-  }
-`;
-
-const Bar = styled.span<{ remainPercent: number }>`
-  position: absolute;
-  border-radius: 46px;
-  background-color: #fce44c;
-  height: 1.481481vh;
-  width: ${(props) => props.remainPercent}%;
-  -webkit-animation: file-progressbar 2s ease-out;
-  animation: file-progressbar 2s ease-out;
-
-  @keyframes file-progressbar {
-    0% {
-      width: 0px;
-    }
-
-    100% {
-      width: ${(props) => props.remainPercent}%;
-    }
-  }
-`;
 
 export default FileInfo;
