+++ src/components/teampleHomePage/SummaryTeample.tsx
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 import progress0 from '../images/progressbar/LoadingIcon_Start.png';
 import progress1 from '../images/progressbar/LoadingIcon_Fire.png';
 import progress2 from '../images/progressbar/LoadingIcon_Turtle.png';
@@ -9,6 +8,7 @@ import axios from 'axios';
 import { useRecoilState } from 'recoil';
 import { teamidState, isLoading5State } from 'state';
 import { baseURL } from 'api/client';
+import * as Style from '../../css/TeampleHomePage/SummaryTeampleStyle';
 
 const SummaryTeample = () => {
   const now = new Date();
@@ -86,14 +86,14 @@ const SummaryTeample = () => {
   }, [doneNum, allNum]);
 
   return (
-    <SummaryContainer>
-      <DateContainer>
+    <Style.SummaryContainer>
+      <Style.DateContainer>
         {year}년 {month}월 {date}일 ({week})
-      </DateContainer>
-      <RemainContainer>
-        <RemainBox>
-          <Text>{text}</Text>
-          <Big>
+      </Style.DateContainer>
+      <Style.RemainContainer>
+        <Style.RemainBox>
+          <Style.Text>{text}</Style.Text>
+          <Style.Big>
             <div style={{ marginTop: '1.6666vh' }}>
               <span>팀플이</span>
               <span style={{ color: '#487AFF', fontWeight: '700' }}>
@@ -104,122 +104,25 @@ const SummaryTeample = () => {
                 <br />
               </span>
             </div>
-          </Big>
-          <Percent>
+          </Style.Big>
+          <Style.Percent>
             ({doneNum}/{allNum})
-          </Percent>
-        </RemainBox>
-      </RemainContainer>
-      <BarContainer>
+          </Style.Percent>
+        </Style.RemainBox>
+      </Style.RemainContainer>
+      <Style.BarContainer>
         <ul>
           <li>
-            <Bar
+            <Style.Bar
               className="teample-progressbar"
               currentPercent={currentPercent}
               icon={icon}
             />
           </li>
         </ul>
-      </BarContainer>
-    </SummaryContainer>
+      </Style.BarContainer>
+    </Style.SummaryContainer>
   );
 };
-
-const SummaryContainer = styled.div`
-  width: 44.2708vw;
-  height: 21.111vh;
-  position: relative;
-`;
-
-const DateContainer = styled.div`
-  position: absolute;
-  left: 2.8125vw;
-  top: 36px;
-  font-size: 0.8333vw;
-  line-height: 100%;
-`;
-
-const RemainContainer = styled.div``;
-
-const RemainBox = styled.div`
-  position: absolute;
-  top: 6.851852vh;
-  left: 2.8125vw;
-`;
-
-const Big = styled.div`
-  font-size: 1.6666vw;
-  line-height: 100%;
-  font-weight: 500;
-`;
-
-const Text = styled.span`
-  font-size: 0.9375vw;
-  font-weight: 400;
-  color: #707070;
-`;
-
-const Percent = styled.div`
-  position: absolute;
-  left: 35.5vw;
-  top: 8.7962vh;
-  line-height: 100%;
-  font-size: 0.9375vw;
-  letter-spacing: 4px;
-  font-weight: 400;
-`;
-
-const BarContainer = styled.div`
-  position: absolute;
-  top: 17.4074vh;
-  left: 2.8125vw;
-  ul {
-    position: relative;
-    padding: 0;
-    list-style: none;
-    width: 40.20833vw;
-  }
-
-  li {
-    background-color: #ececec;
-    height: 13px;
-    border-radius: 46px;
-  }
-`;
-
-const Bar = styled.span<{ currentPercent: number; icon: string }>`
-  position: absolute;
-  border-radius: 46px;
-  background-color: #487aff;
-  height: 13px;
-  width: ${(props) => props.currentPercent}%;
-  -webkit-animation: teample-progressbar 2s ease-out;
-  animation: teample-progressbar 2s ease-out;
-
-  &::after {
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-    content: '';
-    width: 30px;
-    height: 30px;
-    position: absolute;
-    top: -0.92592vh;
-    right: -0.9vw;
-    border-radius: 54px;
-    background-image: url(${(props) => props.icon});
-    background-size: contain;
-    background-repeat: no-repeat;
-  }
-
-  @keyframes teample-progressbar {
-    0% {
-      width: 0px;
-    }
-
-    100% {
-      width: ${(props) => props.currentPercent}%;
-    }
-  }
-`;
 
 export default SummaryTeample;
