+++ src/css/LoginPage/KakaoBtnStyle.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const LoginBtn = styled.div`
+  width: 23vw;
+  height: 6vh;
+  background-color: #fae100;
+  border-radius: 8px;
+  position: relative;
+  display: flex;
+  align-items: center;
+  svg {
+    margin-right: 5vh;
+    position: absolute;
+    left: 2vw;
+  }
+`;
+
+export const Text = styled.span`
+  font-size: 1.2vw;
+  line-height: 100%;
+  color: #383838;
+  position: absolute;
+  left: 6vw;
+  font-weight: 600;
+`;
