+++ src/pages/FilePage.tsx
@@ -3,13 +3,11 @@ import FileHeader from 'components/filePage/FileHeader';
 import Layout from 'components/layouts/layout';
 import FileList from 'components/filePage/fileList';
 
-
 const FilePage = () => {
-
   return (
     <Layout>
-      <FileHeader/>
-      <FileList/>
+      <FileHeader />
+      <FileList />
     </Layout>
   );
 };
