+++ src/pages/ProfilePage.tsx
@@ -4,11 +4,10 @@ import ProfileLayout from 'components/layouts/profileLayout';
 
 const ProfilePage = () => {
   return (
-      <ProfileLayout>
-        <EditProfile />
-      </ProfileLayout>
+    <ProfileLayout>
+      <EditProfile />
+    </ProfileLayout>
   );
 };
-
 
 export default ProfilePage;
