+++ src/pages/TeampleDetailPage.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
-import styled from 'styled-components';
 import Layout from 'components/layouts/layout';
 import DetailBox from 'components/teampleDetailPage/detailBox';
-
-
 
 const TeampleDetailPage = () => {
   return (
     <Layout>
-        <DetailBox/>
+      <DetailBox />
     </Layout>
   );
 };
